Fix keyread_time() argument type
`ranges` is ulong is correct.

https://github.com/MariaDB/server/blob/b66cdbd1eaeed7e96317a03a190c496fd062ec71/storage/mroonga/ha_mroonga.cpp#L13077-L13121

+++ ha_mroonga.cpp
@@ -14987,7 +14987,7 @@ IO_AND_CPU_COST ha_mroonga::rnd_pos_time(ha_rows rows)
 }
 
 IO_AND_CPU_COST ha_mroonga::wrapper_keyread_time(uint index,
-                                                 uint ranges,
+                                                 ulong ranges,
                                                  ha_rows rows,
                                                  ulonglong blocks)
 {
@@ -15018,7 +15018,7 @@ IO_AND_CPU_COST ha_mroonga::wrapper_keyread_time(uint index,
 }
 
 IO_AND_CPU_COST ha_mroonga::storage_keyread_time(uint index,
-                                                 uint ranges,
+                                                 ulong ranges,
                                                  ha_rows rows,
                                                  ulonglong blocks)
 {
@@ -15028,7 +15028,7 @@ IO_AND_CPU_COST ha_mroonga::storage_keyread_time(uint index,
 }
 
 IO_AND_CPU_COST ha_mroonga::keyread_time(uint index,
-                                         uint ranges,
+                                         ulong ranges,
                                          ha_rows rows,
                                          ulonglong blocks)
 {
